Fix: front end Redux toward view all wigs
