+++ client/store/index.js
@@ -1,15 +1,12 @@
-import { createStore, combineReducers, applyMiddleware } from 'redux';
+import { createStore, applyMiddleware } from 'redux';
+import { composeWithDevTools } from 'redux-devtools-extension';
+import rootReducer from './reducers';
+
 import { createLogger } from 'redux-logger';
 import thunkMiddleware from 'redux-thunk';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import user from './user';
-import wigs from './wigs';
 
-const reducer = combineReducers({ user, wigs });
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({ collapsed: true }))
 );
-const store = createStore(reducer, middleware);
-
-export default { store };
-export * from './user';
+export default createStore(rootReducer, middleware);
+export * from './reducers/user';
