+++ client/store/reducers/wigs.spec.js
